+++ packages/sanity/src/core/releases/tool/detail/__tests__/ReleaseTypePicker.test.tsx
@@ -76,7 +76,7 @@ describe('ReleaseTypePicker', () => {
 
       expect(screen.getByTestId('published-release-type-label')).toBeInTheDocument()
 
-      expect(screen.getByText('Published')).toBeInTheDocument()
+      expect(screen.getByText('Published on Oct 10, 2023, 3:00:00 AM')).toBeInTheDocument()
     })
 
     it('renders the label with a published text when release was schedule published', async () => {
